+++ src/main/java/net/silentchaos512/gear/util/GearData.java
@@ -772,10 +772,14 @@ public final class GearData {
         CompoundNBT tags = getData(stack, NBT_ROOT_CONSTRUCTION);
         ListNBT tagList = new ListNBT();
 
-        for (IPartData part : parts) {
-            CompoundNBT write = part.write(new CompoundNBT());
-            tagList.add(write);
-        }
+        // Mains must be first in the list!
+        parts.stream().filter(p -> p.getType() == PartType.MAIN)
+                .map(p -> p.write(new CompoundNBT()))
+                .forEach(tagList::add);
+        // Write everything else in any order
+        parts.stream().filter(p -> p.getType() != PartType.MAIN)
+                .map(p -> p.write(new CompoundNBT()))
+                .forEach(tagList::add);
 
         tags.put(NBT_CONSTRUCTION_PARTS, tagList);
     }